requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
fastapi==0.110.0
google-auth==2.49.1
google-genai==0.3.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==11.3.0
psutil==5.9.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0b2
pydantic_core==2.42.0
Pygments==2.19.2
PyJWT==2.12.1
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
SQLAlchemy==2.0.48
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.1
websockets==14.2
